fix(add-missing-ads-get-by-id-method-move-related-ads-dto-to-one-file): [skp-58]
