+++ src/app/common/dto/advert.dto.ts
