+++ src/app/helpers/schemas/advert.schema.ts
