+++ src/app/view/adverts/adverts.controller.ts
@@ -3,6 +3,7 @@ import {
   HttpStatus,
   HttpCode,
   Get,
+  Param,
 } from '@nestjs/common';
 
 import { AdvertsService } from './adverts.service';
@@ -12,9 +13,16 @@ import { Advert } from 'src/app/common/interfaces/advert.interface';
 export class AdvertsController {
   constructor(private _advertsService: AdvertsService) {}
 
+  @Get('/:id')
+  @HttpCode(HttpStatus.OK)
+  async getAdvertById(@Param('id') id: string): Promise<Advert> {
+    return this._advertsService.getAdvertById(id);
+  }
+
   @Get()
   @HttpCode(HttpStatus.OK)
-  async createAdvert(): Promise<Advert[]> {
+  async getAdverts(): Promise<Advert[]> {
     return this._advertsService.getAdverts();
   }
+  
 }
